@@ -1,53 +1,51 @@
 import 'package:dio/dio.dart';
-import 'package:openapi/api.dart' as openapi;
 
+/// APIClient sınıfı, HTTP istekleri için yapılandırılmış bir Dio istemcisi sağlar.
 class APIClient {
-  final openapi.ApiClient _client;
-  late final openapi.PetApi _petApi;
-  late final openapi.StoreApi _storeApi;
-  late final openapi.UserApi _userApi;
+  final Dio _dio;
 
-  APIClient()
-      : _client =
-            openapi.ApiClient(basePath: 'https://petstore3.swagger.io/api/v3') {
-    _petApi = openapi.PetApi(_client);
-    _storeApi = openapi.StoreApi(_client);
-    _userApi = openapi.UserApi(_client);
+  // API endpoint'leri için temel URL
+  static const String baseUrl = 'https://petstore3.swagger.io/api/v3';
+
+  /// Dio istemcisini dışarı açmak için getter
+  Dio get dio => _dio;
+
+  /// APIClient yapıcı metodu.
+  /// İsteğe bağlı bir Dio örneği alabilir veya varsayılan olarak yeni bir tane oluşturabilir.
+  APIClient({Dio? dio}) : _dio = dio ?? Dio() {
+    // Dio istemcisini yapılandır
+    _dio.options.baseUrl = baseUrl;
+    _dio.options.connectTimeout = const Duration(seconds: 10);
+    _dio.options.receiveTimeout = const Duration(seconds: 10);
+    _dio.options.headers = {
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+    };
+
+    // İstek/yanıt yakalamak için interceptor ekle
+    _dio.interceptors.add(
+      LogInterceptor(requestBody: true, responseBody: true),
+    );
   }
 
-  openapi.PetApi get petApi => _petApi;
-  openapi.StoreApi get storeApi => _storeApi;
-  openapi.UserApi get userApi => _userApi;
-
-  // Eski metodları yeni API sınıflarını kullanacak şekilde güncelle
-  Future<Map<String, dynamic>> getPetById(int id) async {
-    final pet = await _petApi.getPetById(id);
-    return pet?.toJson() ?? {};
-  }
-
-  Future<Response> request({
+  /// Temel istek metodu
+  Future<Response<dynamic>> request({
     required String path,
     String method = 'GET',
     Map<String, dynamic>? queryParameters,
     dynamic data,
-    Map<String, dynamic>? headers,
+    Map<String, String>? headers,
   }) async {
-    final dio = Dio()..options.baseUrl = 'https://petstore3.swagger.io/api/v3';
-    dio.options.headers = {
-      'Content-Type': 'application/json',
-      'Accept': 'application/json',
-      ...?headers,
-    };
-
     try {
-      final response = await dio.request(
+      final response = await _dio.request(
         path,
-        options: Options(method: method),
+        options: Options(method: method, headers: headers),
         queryParameters: queryParameters,
         data: data,
       );
       return response;
     } on DioException catch (e) {
+      // Hata işleme
       print('API İsteği Hatası: ${e.message}');
       rethrow;
     }
