@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../api/api_client.dart';
-import 'package:openapi/api.dart' as openapi;
+import '../models/pet.dart';
 import '../services/pet_service.dart';
 import '../widgets/loading_indicator.dart';
 
@@ -42,12 +42,15 @@ class _AddPetScreenState extends State<AddPetScreen> {
       });
 
       try {
-        final pet = await _petService.addPet(_nameController.text, _status);
-        if (pet != null) {
-          _handleSuccess();
-        } else {
-          _handleError('Pet eklenemedi');
-        }
+        // Yeni bir Pet nesnesi oluştur
+        final pet = Pet(
+          name: _nameController.text,
+          status: _status,
+          photoUrls: [],
+        );
+
+        await _petService.addPet(pet);
+        _handleSuccess();
       } catch (e) {
         _handleError(e.toString());
       } finally {
